added frontend UI

+++ backend/controllers/goalController.js
@@ -1,12 +1,13 @@
 const asyncHandler = require("express-async-handler");
 
 const Goal = require("../model/goalModel");
+const User = require("../model/userModel");
 
 //@desc     GET goals
 //@route    api/goals
 //@access   private
 const getGoals = asyncHandler(async (request, response) => {
-  const goals = await Goal.find();
+  const goals = await Goal.find({ user: request.user.id });
   response.status(200).json(goals);
 });
 
@@ -20,6 +21,7 @@ const setGoals = asyncHandler(async (request, response) => {
   }
   const goal = await Goal.create({
     text: request.body.text,
+    user: request.user.id,
   });
 
   console.log(request.body);
@@ -35,6 +37,19 @@ const updateGoal = asyncHandler(async (request, response) => {
   if (!goal) {
     response.status(400);
     throw new Error("Goal not found");
+  }
+
+  const user = await User.findById(request.user.id);
+
+  //check for user
+  if (!user) {
+    response.status(401);
+    throw new Error("user not found");
+  }
+  //Make sure the logged in user matches the goal user
+  if (goal.user.toString() !== user.id) {
+    response.status(401);
+    throw new Error("User not Authorized");
   }
 
   const updatedGoal = await Goal.findByIdAndUpdate(
@@ -57,6 +72,19 @@ const deleteGoal = asyncHandler(async (request, response) => {
   if (!goal) {
     response.status(400);
     throw new Error("Goal not found");
+  }
+
+  const user = await User.findById(request.user.id);
+
+  //check for user
+  if (!user) {
+    response.status(401);
+    throw new Error("user not found");
+  }
+  //Make sure the logged in user matches the goal user
+  if (goal.user.toString() !== user.id) {
+    response.status(401);
+    throw new Error("User not Authorized");
   }
 
   await Goal.findByIdAndDelete(request.params.id);
